refactor(qd): split quad-double into core / manipulators / iostream (#1334)
    core.hpp          arithmetic, no streams          73,741 lines, 0 I/O-family
    manipulators.hpp  type_tag / to_quad / to_triple / to_binary / to_native / to_components
    iostream.hpp      operator<< / operator>>
    qd.hpp            everything, as before

The umbrella was 86,857 lines pulling all five stream headers.

Same division as dd (#1426), for the same reasons: to_string() and parse() stay in the
CORE. parse() is what assign(const std::string&) calls, and to_string() concatenates a
std::string without ever opening a stream -- it only takes std::streamsize and
std::ios_base flags, which is what <ios> is for. What moved to manipulators.hpp is the
contiguous run of five free builders that genuinely use <sstream>: to_quad, to_triple,
to_binary, to_native and to_components, which sat between ulp() and reciprocal() in the
middle of the arithmetic.

qd has no bTrace* prints, so unlike dd the only diagnostics were two std::cerr calls in
to_digits(); both now use fprintf. qd_impl.hpp took the full native/ieee754.hpp for
scale(), which is bit manipulation, and is re-pointed at ieee754_core.hpp and
manipulators_core.hpp.

qd/manipulators.hpp had no #pragma once -- the same defect dd's had, and for the same
reason it mattered: attributes.hpp now includes it as well as the umbrella. That was one
of the 9 headers tree-wide lacking a guard; with dd's fixed in #1426 this leaves 7.

qd_fwd.hpp was not self-contained: it declares parse(const std::string&, qd&) and included
only <cstdint>, so it failed on its own with 'string in namespace std does not name a
type'. dd_fwd.hpp compiles clean, which is why the asymmetry was worth checking rather
than assuming both were fine.

IWYU sweep clean at 0 gaps across all 11 qd headers, including <string> in exceptions.hpp
and <algorithm> for the std::max in to_string()'s fixed-format path.

Verified: the core gate passes on gcc and clang, with #error tripwires on the stream
include guards and exact dyadic identities so the assertions do not depend on
quad-double rounding; static/highprecision -- dd, qd AND the three cascades -- 115 pass
0 fail on both compilers; every api.cpp in the tree 55 pass 0 fail on both;
static/appenv/multifile, which links dd.cpp and qd.cpp into one program, links and runs
clean; ASCII clean.
