+++ include/sw/universal/number/qd/qd.hpp
@@ -13,8 +13,6 @@
 
 ////////////////////////////////////////////////////////////////////////////////////////
 /// required std libraries 
-#include <iostream>
-#include <iomanip>
 
 ////////////////////////////////////////////////////////////////////////////////////////
 ///  BEHAVIORAL COMPILATION SWITCHES
@@ -58,24 +56,26 @@
 
 ///////////////////////////////////////////////////////////////////////////////////////
 // bring in the trait functions
-#include <universal/traits/number_traits.hpp>
+/// INCLUDE FILES that make up the library
+///
+/// Layered per #1334:
+///     core.hpp          arithmetic, no streams
+///     manipulators.hpp  type_tag / to_quad / to_triple / to_binary / to_native / to_components
+///     iostream.hpp      operator<< / operator>>
+///
+/// A translation unit that only computes can include core.hpp directly and skip the
+/// stream headers entirely. to_string() and parse() are in the CORE, not the text
+/// layers: assign(const std::string&) calls parse(), and to_string() concatenates a
+/// std::string without ever touching a stream.
+#include <universal/number/qd/core.hpp>
+#include <universal/number/qd/manipulators.hpp>
+#include <universal/number/qd/iostream.hpp>
+#include <universal/number/qd/attributes.hpp>
+
+/// the report builders stay at the umbrella: they pull <sstream>/<iomanip> by design
 #include <universal/traits/arithmetic_traits.hpp>
 #include <universal/common/number_traits_reports.hpp>
 
-////////////////////////////////////////////////////////////////////////////////////////
-/// INCLUDE FILES that make up the library
-#include <universal/number/qd/exceptions.hpp>
-#include <universal/number/qd/qd_fwd.hpp>
-#include <universal/number/qd/qd_impl.hpp>
-#include <universal/traits/qd_traits.hpp>
-#include <universal/number/qd/numeric_limits.hpp>
-
-// useful functions to work with doubledoubles
-#include <universal/number/qd/manipulators.hpp>
-#include <universal/number/qd/attributes.hpp>
-
-///////////////////////////////////////////////////////////////////////////////////////
-/// elementary math functions library
 #include <universal/number/qd/math/constants/qd_constants.hpp>
 #include <universal/number/qd/mathlib.hpp>
 #include <universal/number/qd/mathext.hpp>
