+++ include/sw/universal/number/qd/qd_fwd.hpp
@@ -6,6 +6,7 @@
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
 #include <cstdint>
+#include <string>   // std::string, in the parse() declaration below
 
 namespace sw { namespace universal {
